MODORDERS-685 	Implement edit instance connection logic of POL - create inventory set to Instance or none
